+++ examples/tag_reader.rs
@@ -33,8 +33,7 @@ fn main() {
 	// import keys from https://docs.rs/lofty/latest/lofty/enum.ItemKey.html
 	println!(
 		"Album Artist: {}",
-		tag.get_string(&lofty::ItemKey::AlbumArtist)
-			.unwrap_or("None")
+		tag.get_string(&ItemKey::AlbumArtist).unwrap_or("None")
 	);
 
 	let properties = tagged_file.properties();